Read CLI "files" with importer
Use the jsonnet importer to read the "filenames" given on the
command line. This allows the filenames to be URLs so that kubecfg can
be run directly on a remote URL instead of needing a local file.

This adds some level of parity to kubectl which can read manifests from
URLs given on the command line.